+++ ChangeLog.md
@@ -11,7 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### added
 
-- added ...
+- added dynamic (thread-local) binding. e.g: `(binding [x 100] ...)
 
 ### fixed